Initial Commit of Anarkey
I've made a basic proof of concept of the anarkey generic key
encrypting method. I am aware that this is basically just the basic
way of doing RSA+AES, but I'm new to crypto so I'm feeling pretty
good about this. I need to proove out the keychain concept next.
That is when Anarkey will really start to take shape into what it
is supposed to be.