Initial work on CLI program.
I'm starting to work on a commandline program to be the simplest
possible form of the program I can make that might be some what
useful. In doing so I'm being forced to flesh out a number of my
concepts to make them actually functional. I'm trying to contain my
self and not bog myself down it trying to foresee all possible
obstacles. It's going well so far.